requirements.txt
numpy>=1.24
torch>=2.0
PyYAML>=6.0
pygame>=2.5
